+++ sdk/resourcemanager/baremetalinfrastructure/armbaremetalinfrastructure/azurebaremetalinstances_client.go
@@ -43,10 +43,134 @@ func NewAzureBareMetalInstancesClient(subscriptionID string, credential azcore.T
 	return client, nil
 }
 
+// Create - Creates an Azure Bare Metal Instance for the specified subscription, resource group, and instance name.
+// If the operation fails it returns an *azcore.ResponseError type.
+//
+// Generated from API version 2024-08-01-preview
+//   - resourceGroupName - The name of the resource group. The name is case insensitive.
+//   - azureBareMetalInstanceName - Name of the Azure Bare Metal Instance, also known as the ResourceName.
+//   - requestBodyParameters - Resource create parameters.
+//   - options - AzureBareMetalInstancesClientCreateOptions contains the optional parameters for the AzureBareMetalInstancesClient.Create
+//     method.
+func (client *AzureBareMetalInstancesClient) Create(ctx context.Context, resourceGroupName string, azureBareMetalInstanceName string, requestBodyParameters AzureBareMetalInstance, options *AzureBareMetalInstancesClientCreateOptions) (AzureBareMetalInstancesClientCreateResponse, error) {
+	var err error
+	const operationName = "AzureBareMetalInstancesClient.Create"
+	ctx = context.WithValue(ctx, runtime.CtxAPINameKey{}, operationName)
+	ctx, endSpan := runtime.StartSpan(ctx, operationName, client.internal.Tracer(), nil)
+	defer func() { endSpan(err) }()
+	req, err := client.createCreateRequest(ctx, resourceGroupName, azureBareMetalInstanceName, requestBodyParameters, options)
+	if err != nil {
+		return AzureBareMetalInstancesClientCreateResponse{}, err
+	}
+	httpResp, err := client.internal.Pipeline().Do(req)
+	if err != nil {
+		return AzureBareMetalInstancesClientCreateResponse{}, err
+	}
+	if !runtime.HasStatusCode(httpResp, http.StatusOK, http.StatusCreated) {
+		err = runtime.NewResponseError(httpResp)
+		return AzureBareMetalInstancesClientCreateResponse{}, err
+	}
+	resp, err := client.createHandleResponse(httpResp)
+	return resp, err
+}
+
+// createCreateRequest creates the Create request.
+func (client *AzureBareMetalInstancesClient) createCreateRequest(ctx context.Context, resourceGroupName string, azureBareMetalInstanceName string, requestBodyParameters AzureBareMetalInstance, options *AzureBareMetalInstancesClientCreateOptions) (*policy.Request, error) {
+	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.BareMetalInfrastructure/bareMetalInstances/{azureBareMetalInstanceName}"
+	if client.subscriptionID == "" {
+		return nil, errors.New("parameter client.subscriptionID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{subscriptionId}", url.PathEscape(client.subscriptionID))
+	if resourceGroupName == "" {
+		return nil, errors.New("parameter resourceGroupName cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{resourceGroupName}", url.PathEscape(resourceGroupName))
+	if azureBareMetalInstanceName == "" {
+		return nil, errors.New("parameter azureBareMetalInstanceName cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{azureBareMetalInstanceName}", url.PathEscape(azureBareMetalInstanceName))
+	req, err := runtime.NewRequest(ctx, http.MethodPut, runtime.JoinPaths(client.internal.Endpoint(), urlPath))
+	if err != nil {
+		return nil, err
+	}
+	reqQP := req.Raw().URL.Query()
+	reqQP.Set("api-version", "2024-08-01-preview")
+	req.Raw().URL.RawQuery = reqQP.Encode()
+	req.Raw().Header["Accept"] = []string{"application/json"}
+	if err := runtime.MarshalAsJSON(req, requestBodyParameters); err != nil {
+		return nil, err
+	}
+	return req, nil
+}
+
+// createHandleResponse handles the Create response.
+func (client *AzureBareMetalInstancesClient) createHandleResponse(resp *http.Response) (AzureBareMetalInstancesClientCreateResponse, error) {
+	result := AzureBareMetalInstancesClientCreateResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.AzureBareMetalInstance); err != nil {
+		return AzureBareMetalInstancesClientCreateResponse{}, err
+	}
+	return result, nil
+}
+
+// Delete - Deletes an Azure Bare Metal Instance for the specified subscription, resource group, and instance name.
+// If the operation fails it returns an *azcore.ResponseError type.
+//
+// Generated from API version 2024-08-01-preview
+//   - resourceGroupName - The name of the resource group. The name is case insensitive.
+//   - azureBareMetalInstanceName - Name of the Azure Bare Metal Instance, also known as the ResourceName.
+//   - options - AzureBareMetalInstancesClientDeleteOptions contains the optional parameters for the AzureBareMetalInstancesClient.Delete
+//     method.
+func (client *AzureBareMetalInstancesClient) Delete(ctx context.Context, resourceGroupName string, azureBareMetalInstanceName string, options *AzureBareMetalInstancesClientDeleteOptions) (AzureBareMetalInstancesClientDeleteResponse, error) {
+	var err error
+	const operationName = "AzureBareMetalInstancesClient.Delete"
+	ctx = context.WithValue(ctx, runtime.CtxAPINameKey{}, operationName)
+	ctx, endSpan := runtime.StartSpan(ctx, operationName, client.internal.Tracer(), nil)
+	defer func() { endSpan(err) }()
+	req, err := client.deleteCreateRequest(ctx, resourceGroupName, azureBareMetalInstanceName, options)
+	if err != nil {
+		return AzureBareMetalInstancesClientDeleteResponse{}, err
+	}
+	httpResp, err := client.internal.Pipeline().Do(req)
+	if err != nil {
+		return AzureBareMetalInstancesClientDeleteResponse{}, err
+	}
+	if !runtime.HasStatusCode(httpResp, http.StatusOK, http.StatusNoContent) {
+		err = runtime.NewResponseError(httpResp)
+		return AzureBareMetalInstancesClientDeleteResponse{}, err
+	}
+	return AzureBareMetalInstancesClientDeleteResponse{}, nil
+}
+
+// deleteCreateRequest creates the Delete request.
+func (client *AzureBareMetalInstancesClient) deleteCreateRequest(ctx context.Context, resourceGroupName string, azureBareMetalInstanceName string, options *AzureBareMetalInstancesClientDeleteOptions) (*policy.Request, error) {
+	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.BareMetalInfrastructure/bareMetalInstances/{azureBareMetalInstanceName}"
+	if client.subscriptionID == "" {
+		return nil, errors.New("parameter client.subscriptionID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{subscriptionId}", url.PathEscape(client.subscriptionID))
+	if resourceGroupName == "" {
+		return nil, errors.New("parameter resourceGroupName cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{resourceGroupName}", url.PathEscape(resourceGroupName))
+	if azureBareMetalInstanceName == "" {
+		return nil, errors.New("parameter azureBareMetalInstanceName cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{azureBareMetalInstanceName}", url.PathEscape(azureBareMetalInstanceName))
+	req, err := runtime.NewRequest(ctx, http.MethodDelete, runtime.JoinPaths(client.internal.Endpoint(), urlPath))
+	if err != nil {
+		return nil, err
+	}
+	reqQP := req.Raw().URL.Query()
+	reqQP.Set("api-version", "2024-08-01-preview")
+	req.Raw().URL.RawQuery = reqQP.Encode()
+	req.Raw().Header["Accept"] = []string{"application/json"}
+	return req, nil
+}
+
 // Get - Gets an Azure Bare Metal Instance for the specified subscription, resource group, and instance name.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-08-04-preview
+// Generated from API version 2024-08-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - azureBareMetalInstanceName - Name of the Azure Bare Metal Instance, also known as the ResourceName.
 //   - options - AzureBareMetalInstancesClientGetOptions contains the optional parameters for the AzureBareMetalInstancesClient.Get
@@ -93,7 +217,7 @@ func (client *AzureBareMetalInstancesClient) getCreateRequest(ctx context.Contex
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-08-04-preview")
+	reqQP.Set("api-version", "2024-08-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -111,7 +235,7 @@ func (client *AzureBareMetalInstancesClient) getHandleResponse(resp *http.Respon
 // NewListByResourceGroupPager - Gets a list of Azure Bare Metal Instances in the specified subscription and resource group.
 // The operations returns various properties of each Azure Bare Metal Instance.
 //
-// Generated from API version 2023-08-04-preview
+// Generated from API version 2024-08-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - options - AzureBareMetalInstancesClientListByResourceGroupOptions contains the optional parameters for the AzureBareMetalInstancesClient.NewListByResourceGroupPager
 //     method.
@@ -154,7 +278,7 @@ func (client *AzureBareMetalInstancesClient) listByResourceGroupCreateRequest(ct
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-08-04-preview")
+	reqQP.Set("api-version", "2024-08-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -172,7 +296,7 @@ func (client *AzureBareMetalInstancesClient) listByResourceGroupHandleResponse(r
 // NewListBySubscriptionPager - Returns a list of Azure Bare Metal Instances in the specified subscription. The operations
 // returns various properties of each Azure Bare Metal Instance.
 //
-// Generated from API version 2023-08-04-preview
+// Generated from API version 2024-08-01-preview
 //   - options - AzureBareMetalInstancesClientListBySubscriptionOptions contains the optional parameters for the AzureBareMetalInstancesClient.NewListBySubscriptionPager
 //     method.
 func (client *AzureBareMetalInstancesClient) NewListBySubscriptionPager(options *AzureBareMetalInstancesClientListBySubscriptionOptions) *runtime.Pager[AzureBareMetalInstancesClientListBySubscriptionResponse] {
@@ -210,7 +334,7 @@ func (client *AzureBareMetalInstancesClient) listBySubscriptionCreateRequest(ctx
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-08-04-preview")
+	reqQP.Set("api-version", "2024-08-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -228,7 +352,7 @@ func (client *AzureBareMetalInstancesClient) listBySubscriptionHandleResponse(re
 // BeginRestart - The operation to restart an Azure Bare Metal Instance
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-08-04-preview
+// Generated from API version 2024-08-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - azureBareMetalInstanceName - Name of the Azure Bare Metal Instance, also known as the ResourceName.
 //   - options - AzureBareMetalInstancesClientBeginRestartOptions contains the optional parameters for the AzureBareMetalInstancesClient.BeginRestart
@@ -254,7 +378,7 @@ func (client *AzureBareMetalInstancesClient) BeginRestart(ctx context.Context, r
 // Restart - The operation to restart an Azure Bare Metal Instance
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-08-04-preview
+// Generated from API version 2024-08-01-preview
 func (client *AzureBareMetalInstancesClient) restart(ctx context.Context, resourceGroupName string, azureBareMetalInstanceName string, options *AzureBareMetalInstancesClientBeginRestartOptions) (*http.Response, error) {
 	var err error
 	const operationName = "AzureBareMetalInstancesClient.BeginRestart"
@@ -296,7 +420,7 @@ func (client *AzureBareMetalInstancesClient) restartCreateRequest(ctx context.Co
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-08-04-preview")
+	reqQP.Set("api-version", "2024-08-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if options != nil && options.ForceParameter != nil {
@@ -311,7 +435,7 @@ func (client *AzureBareMetalInstancesClient) restartCreateRequest(ctx context.Co
 // BeginShutdown - The operation to shutdown an Azure Bare Metal Instance
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-08-04-preview
+// Generated from API version 2024-08-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - azureBareMetalInstanceName - Name of the Azure Bare Metal Instance, also known as the ResourceName.
 //   - options - AzureBareMetalInstancesClientBeginShutdownOptions contains the optional parameters for the AzureBareMetalInstancesClient.BeginShutdown
@@ -337,7 +461,7 @@ func (client *AzureBareMetalInstancesClient) BeginShutdown(ctx context.Context, 
 // Shutdown - The operation to shutdown an Azure Bare Metal Instance
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-08-04-preview
+// Generated from API version 2024-08-01-preview
 func (client *AzureBareMetalInstancesClient) shutdown(ctx context.Context, resourceGroupName string, azureBareMetalInstanceName string, options *AzureBareMetalInstancesClientBeginShutdownOptions) (*http.Response, error) {
 	var err error
 	const operationName = "AzureBareMetalInstancesClient.BeginShutdown"
@@ -379,7 +503,7 @@ func (client *AzureBareMetalInstancesClient) shutdownCreateRequest(ctx context.C
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-08-04-preview")
+	reqQP.Set("api-version", "2024-08-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -388,7 +512,7 @@ func (client *AzureBareMetalInstancesClient) shutdownCreateRequest(ctx context.C
 // BeginStart - The operation to start an Azure Bare Metal instance
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-08-04-preview
+// Generated from API version 2024-08-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - azureBareMetalInstanceName - Name of the Azure Bare Metal Instance, also known as the ResourceName.
 //   - options - AzureBareMetalInstancesClientBeginStartOptions contains the optional parameters for the AzureBareMetalInstancesClient.BeginStart
@@ -414,7 +538,7 @@ func (client *AzureBareMetalInstancesClient) BeginStart(ctx context.Context, res
 // Start - The operation to start an Azure Bare Metal instance
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-08-04-preview
+// Generated from API version 2024-08-01-preview
 func (client *AzureBareMetalInstancesClient) start(ctx context.Context, resourceGroupName string, azureBareMetalInstanceName string, options *AzureBareMetalInstancesClientBeginStartOptions) (*http.Response, error) {
 	var err error
 	const operationName = "AzureBareMetalInstancesClient.BeginStart"
@@ -456,7 +580,7 @@ func (client *AzureBareMetalInstancesClient) startCreateRequest(ctx context.Cont
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-08-04-preview")
+	reqQP.Set("api-version", "2024-08-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -466,7 +590,7 @@ func (client *AzureBareMetalInstancesClient) startCreateRequest(ctx context.Cont
 // name.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-08-04-preview
+// Generated from API version 2024-08-01-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - azureBareMetalInstanceName - Name of the Azure Bare Metal Instance, also known as the ResourceName.
 //   - tagsParameter - Request body that only contains the new Tags field
@@ -514,7 +638,7 @@ func (client *AzureBareMetalInstancesClient) updateCreateRequest(ctx context.Con
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-08-04-preview")
+	reqQP.Set("api-version", "2024-08-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, tagsParameter); err != nil {