+++ sdk/resourcemanager/baremetalinfrastructure/armbaremetalinfrastructure/fake/azurebaremetalinstances_server.go
@@ -25,6 +25,14 @@ import (
 
 // AzureBareMetalInstancesServer is a fake server for instances of the armbaremetalinfrastructure.AzureBareMetalInstancesClient type.
 type AzureBareMetalInstancesServer struct {
+	// Create is the fake for method AzureBareMetalInstancesClient.Create
+	// HTTP status codes to indicate success: http.StatusOK, http.StatusCreated
+	Create func(ctx context.Context, resourceGroupName string, azureBareMetalInstanceName string, requestBodyParameters armbaremetalinfrastructure.AzureBareMetalInstance, options *armbaremetalinfrastructure.AzureBareMetalInstancesClientCreateOptions) (resp azfake.Responder[armbaremetalinfrastructure.AzureBareMetalInstancesClientCreateResponse], errResp azfake.ErrorResponder)
+
+	// Delete is the fake for method AzureBareMetalInstancesClient.Delete
+	// HTTP status codes to indicate success: http.StatusOK, http.StatusNoContent
+	Delete func(ctx context.Context, resourceGroupName string, azureBareMetalInstanceName string, options *armbaremetalinfrastructure.AzureBareMetalInstancesClientDeleteOptions) (resp azfake.Responder[armbaremetalinfrastructure.AzureBareMetalInstancesClientDeleteResponse], errResp azfake.ErrorResponder)
+
 	// Get is the fake for method AzureBareMetalInstancesClient.Get
 	// HTTP status codes to indicate success: http.StatusOK
 	Get func(ctx context.Context, resourceGroupName string, azureBareMetalInstanceName string, options *armbaremetalinfrastructure.AzureBareMetalInstancesClientGetOptions) (resp azfake.Responder[armbaremetalinfrastructure.AzureBareMetalInstancesClientGetResponse], errResp azfake.ErrorResponder)
@@ -91,6 +99,10 @@ func (a *AzureBareMetalInstancesServerTransport) Do(req *http.Request) (*http.Re
 	var err error
 
 	switch method {
+	case "AzureBareMetalInstancesClient.Create":
+		resp, err = a.dispatchCreate(req)
+	case "AzureBareMetalInstancesClient.Delete":
+		resp, err = a.dispatchDelete(req)
 	case "AzureBareMetalInstancesClient.Get":
 		resp, err = a.dispatchGet(req)
 	case "AzureBareMetalInstancesClient.NewListByResourceGroupPager":
@@ -113,6 +125,76 @@ func (a *AzureBareMetalInstancesServerTransport) Do(req *http.Request) (*http.Re
 		return nil, err
 	}
 
+	return resp, nil
+}
+
+func (a *AzureBareMetalInstancesServerTransport) dispatchCreate(req *http.Request) (*http.Response, error) {
+	if a.srv.Create == nil {
+		return nil, &nonRetriableError{errors.New("fake for method Create not implemented")}
+	}
+	const regexStr = `/subscriptions/(?P<subscriptionId>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/resourceGroups/(?P<resourceGroupName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/providers/Microsoft\.BareMetalInfrastructure/bareMetalInstances/(?P<azureBareMetalInstanceName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)`
+	regex := regexp.MustCompile(regexStr)
+	matches := regex.FindStringSubmatch(req.URL.EscapedPath())
+	if matches == nil || len(matches) < 3 {
+		return nil, fmt.Errorf("failed to parse path %s", req.URL.Path)
+	}
+	body, err := server.UnmarshalRequestAsJSON[armbaremetalinfrastructure.AzureBareMetalInstance](req)
+	if err != nil {
+		return nil, err
+	}
+	resourceGroupNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("resourceGroupName")])
+	if err != nil {
+		return nil, err
+	}
+	azureBareMetalInstanceNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("azureBareMetalInstanceName")])
+	if err != nil {
+		return nil, err
+	}
+	respr, errRespr := a.srv.Create(req.Context(), resourceGroupNameParam, azureBareMetalInstanceNameParam, body, nil)
+	if respErr := server.GetError(errRespr, req); respErr != nil {
+		return nil, respErr
+	}
+	respContent := server.GetResponseContent(respr)
+	if !contains([]int{http.StatusOK, http.StatusCreated}, respContent.HTTPStatus) {
+		return nil, &nonRetriableError{fmt.Errorf("unexpected status code %d. acceptable values are http.StatusOK, http.StatusCreated", respContent.HTTPStatus)}
+	}
+	resp, err := server.MarshalResponseAsJSON(respContent, server.GetResponse(respr).AzureBareMetalInstance, req)
+	if err != nil {
+		return nil, err
+	}
+	return resp, nil
+}
+
+func (a *AzureBareMetalInstancesServerTransport) dispatchDelete(req *http.Request) (*http.Response, error) {
+	if a.srv.Delete == nil {
+		return nil, &nonRetriableError{errors.New("fake for method Delete not implemented")}
+	}
+	const regexStr = `/subscriptions/(?P<subscriptionId>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/resourceGroups/(?P<resourceGroupName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/providers/Microsoft\.BareMetalInfrastructure/bareMetalInstances/(?P<azureBareMetalInstanceName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)`
+	regex := regexp.MustCompile(regexStr)
+	matches := regex.FindStringSubmatch(req.URL.EscapedPath())
+	if matches == nil || len(matches) < 3 {
+		return nil, fmt.Errorf("failed to parse path %s", req.URL.Path)
+	}
+	resourceGroupNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("resourceGroupName")])
+	if err != nil {
+		return nil, err
+	}
+	azureBareMetalInstanceNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("azureBareMetalInstanceName")])
+	if err != nil {
+		return nil, err
+	}
+	respr, errRespr := a.srv.Delete(req.Context(), resourceGroupNameParam, azureBareMetalInstanceNameParam, nil)
+	if respErr := server.GetError(errRespr, req); respErr != nil {
+		return nil, respErr
+	}
+	respContent := server.GetResponseContent(respr)
+	if !contains([]int{http.StatusOK, http.StatusNoContent}, respContent.HTTPStatus) {
+		return nil, &nonRetriableError{fmt.Errorf("unexpected status code %d. acceptable values are http.StatusOK, http.StatusNoContent", respContent.HTTPStatus)}
+	}
+	resp, err := server.NewResponse(respContent, req, nil)
+	if err != nil {
+		return nil, err
+	}
 	return resp, nil
 }