+++ sdk/resourcemanager/baremetalinfrastructure/armbaremetalinfrastructure/models.go
@@ -36,8 +36,14 @@ type AzureBareMetalInstance struct {
 
 // AzureBareMetalInstanceProperties - Describes the properties of an Azure Bare Metal Instance.
 type AzureBareMetalInstanceProperties struct {
+	// Specifies the Azure Bare Metal Instance unique ID.
+	AzureBareMetalInstanceID *string
+
 	// Specifies the hardware settings for the Azure Bare Metal Instance.
 	HardwareProfile *HardwareProfile
+
+	// Hardware revision of an Azure Bare Metal Instance
+	HwRevision *string
 
 	// Specifies the network settings for the Azure Bare Metal Instance.
 	NetworkProfile *NetworkProfile
@@ -48,23 +54,17 @@ type AzureBareMetalInstanceProperties struct {
 	// ARM ID of another AzureBareMetalInstance that will share a network with this AzureBareMetalInstance
 	PartnerNodeID *string
 
+	// Resource power state
+	PowerState *AzureBareMetalInstancePowerStateEnum
+
+	// Resource proximity placement group
+	ProximityPlacementGroup *string
+
 	// Specifies the storage settings for the Azure Bare Metal Instance disks.
 	StorageProfile *StorageProfile
 
-	// READ-ONLY; Specifies the Azure Bare Metal Instance unique ID.
-	AzureBareMetalInstanceID *string
-
-	// READ-ONLY; Hardware revision of an Azure Bare Metal Instance
-	HwRevision *string
-
-	// READ-ONLY; Resource power state
-	PowerState *AzureBareMetalInstancePowerStateEnum
-
 	// READ-ONLY; State of provisioning of the AzureBareMetalInstance
 	ProvisioningState *AzureBareMetalProvisioningStatesEnum
-
-	// READ-ONLY; Resource proximity placement group
-	ProximityPlacementGroup *string
 }
 
 // AzureBareMetalInstancesListResult - The response from the List Azure Bare Metal Instances operation.
@@ -80,6 +80,9 @@ type AzureBareMetalInstancesListResult struct {
 type AzureBareMetalStorageInstance struct {
 	// REQUIRED; The geo-location where the resource lives
 	Location *string
+
+	// The identity of Azure Bare Metal Storage Instance, if configured.
+	Identity *AzureBareMetalStorageInstanceIdentity
 
 	// AzureBareMetalStorageInstance properties
 	Properties *AzureBareMetalStorageInstanceProperties
@@ -98,6 +101,31 @@ type AzureBareMetalStorageInstance struct {
 
 	// READ-ONLY; The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts"
 	Type *string
+}
+
+// AzureBareMetalStorageInstanceBody - properties of body during PUT/PATCH for an AzureBareMetalStorageInstance.
+type AzureBareMetalStorageInstanceBody struct {
+	// The identity of Azure Bare Metal Storage Instance, if configured.
+	Identity *AzureBareMetalStorageInstanceIdentity
+
+	// Tags field of the AzureBareMetal/AzureBareMetaStorage instance.
+	Tags map[string]*string
+}
+
+// AzureBareMetalStorageInstanceIdentity - Identity for Azure Bare Metal Storage Instance.
+type AzureBareMetalStorageInstanceIdentity struct {
+	// The type of identity used for the Azure Bare Metal Storage Instance. The type 'SystemAssigned' refers to an implicitly
+	// created identity. The type 'None' will remove any identities from the Azure Bare
+	// Metal Storage Instance.
+	Type *ResourceIdentityType
+
+	// READ-ONLY; The principal ID of Azure Bare Metal Storage Instance identity. This property will only be provided for a system
+	// assigned identity.
+	PrincipalID *string
+
+	// READ-ONLY; The tenant ID associated with the Azure Bare Metal Storage Instance. This property will only be provided for
+	// a system assigned identity.
+	TenantID *string
 }
 
 // AzureBareMetalStorageInstanceProperties - Describes the properties of an AzureBareMetalStorageInstance.
@@ -140,10 +168,10 @@ type ForceState struct {
 
 // HardwareProfile - Specifies the hardware settings for the Azure Bare Metal Instance.
 type HardwareProfile struct {
-	// READ-ONLY; Specifies the Azure Bare Metal Instance SKU.
+	// Specifies the Azure Bare Metal Instance SKU.
 	AzureBareMetalInstanceSize *AzureBareMetalInstanceSizeNamesEnum
 
-	// READ-ONLY; Name of the hardware type (vendor and/or their product name)
+	// Name of the hardware type (vendor and/or their product name)
 	HardwareType *AzureBareMetalHardwareTypeNamesEnum
 }
 
@@ -155,11 +183,11 @@ type NetworkInterface struct {
 
 // NetworkProfile - Specifies the network settings for the Azure Bare Metal Instance disks.
 type NetworkProfile struct {
+	// Specifies the circuit id for connecting to express route.
+	CircuitID *string
+
 	// Specifies the network interfaces for the Azure Bare Metal Instance.
 	NetworkInterfaces []*NetworkInterface
-
-	// READ-ONLY; Specifies the circuit id for connecting to express route.
-	CircuitID *string
 }
 
 // OSProfile - Specifies the operating system settings for the Azure Bare Metal instance.
@@ -167,13 +195,13 @@ type OSProfile struct {
 	// Specifies the host OS name of the Azure Bare Metal instance.
 	ComputerName *string
 
+	// This property allows you to specify the type of the OS.
+	OSType *string
+
 	// Specifies the SSH public key used to access the operating system.
 	SSHPublicKey *string
 
-	// READ-ONLY; This property allows you to specify the type of the OS.
-	OSType *string
-
-	// READ-ONLY; Specifies version of operating system.
+	// Specifies version of operating system.
 	Version *string
 }
 
@@ -261,11 +289,11 @@ type StorageBillingProperties struct {
 
 // StorageProfile - Specifies the storage settings for the Azure Bare Metal instance disks.
 type StorageProfile struct {
+	// IP Address to connect to storage.
+	NfsIPAddress *string
+
 	// Specifies information about the operating system disk used by bare metal instance.
 	OSDisks []*Disk
-
-	// READ-ONLY; IP Address to connect to storage.
-	NfsIPAddress *string
 }
 
 // StorageProperties - described the storage properties of the azure bare metal storage instance
@@ -313,7 +341,7 @@ type SystemData struct {
 	LastModifiedByType *CreatedByType
 }
 
-// Tags field of the AzureBareMetal/AzureBareMetaStorage instance.
+// Tags field of the AzureBareMetal/AzureBareMetalStorage instance.
 type Tags struct {
 	// Tags field of the AzureBareMetal/AzureBareMetaStorage instance.
 	Tags map[string]*string